Add public enum class for WantedReason
